verificações em todos os endpoints sensíveis e logout

+++ src/main/java/com/oroboros/oroboros/controller/ImagemProdutoController.java
@@ -3,6 +3,7 @@ package com.oroboros.oroboros.controller;
 import java.io.IOException;
 import java.util.List;
 
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -27,46 +28,61 @@ import com.oroboros.oroboros.util.EntidadeException;
 @RequestMapping("/imgprod")
 public class ImagemProdutoController {
 
-    @Autowired
-    private ProdutoRepository pr;
+   @Autowired
+   private ProdutoRepository pr;
 
-    @Autowired
-    private ImagemProdutoRepository ir;
+   @Autowired
+   private ImagemProdutoRepository ir;
+
+   @Autowired
+   UsuarioController uc = new UsuarioController();
 
    @PutMapping("/upload")
-    public ResponseEntity<?> uploadIMGPROD(@RequestParam("image") MultipartFile file, @RequestParam("id") Long id) {
-        Produto p = pr.findById(id).orElseThrow(() -> new EntidadeException("Produto", id));
-        ImagemProduto imgProd = new ImagemProduto();
-        try {
+   public ResponseEntity<?> uploadIMGPROD(@RequestParam("image") MultipartFile file, @RequestParam("id") Long id, HttpServletRequest request) {
+      if (!uc.checkAdmin(request)) {
+         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Acesso negado!");
+      } else {
+         Produto p = pr.findById(id).orElseThrow(() -> new EntidadeException("Produto", id));
+         ImagemProduto imgProd = new ImagemProduto();
+         try {
             imgProd.setProduto(p);
             imgProd.setDados(file.getBytes());
             ir.save(imgProd);
-        } catch (IOException e) {
+         } catch (IOException e) {
             ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Erro ao salvar imagem!");
-        }
-        return ResponseEntity.status(HttpStatus.OK).body("Imagem salva com sucesso!");
-    }
+         }
+         return ResponseEntity.status(HttpStatus.OK).body("Imagem salva com sucesso!");
+      }
+   }
 
    @PostMapping("/upload/{id}")
-    public ResponseEntity<?> atualiza(@PathVariable Long id, @RequestBody ImagemProduto a) { 
-        Produto p = pr.findById(id).orElseThrow(() -> new EntidadeException("Produto", id));
-        a.setProduto(p);
-        ir.save(a);
-        return ResponseEntity.status(HttpStatus.OK).body("Imagem salva com sucesso!");
-    }
+   public ResponseEntity<?> atualiza(@PathVariable Long id, @RequestBody ImagemProduto a, HttpServletRequest request) {
+      if (!uc.checkAdmin(request)) {
+         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Acesso negado!");
+      } else {
+         Produto p = pr.findById(id).orElseThrow(() -> new EntidadeException("Produto", id));
+         a.setProduto(p);
+         ir.save(a);
+         return ResponseEntity.status(HttpStatus.OK).body("Imagem salva com sucesso!");
+      }
+   }
 
 
-    @GetMapping("/prodid/{id}")
-    public List<ImagemProduto> getIMGPROD(@PathVariable Long id) {
-        Produto p = pr.findById(id).orElseThrow(() -> new EntidadeException("Produto", id));
-        return ir.findByProduto(p);
-    }
+   @GetMapping("/prodid/{id}")
+   public List<ImagemProduto> getIMGPROD(@PathVariable Long id) {
+      Produto p = pr.findById(id).orElseThrow(() -> new EntidadeException("Produto", id));
+      return ir.findByProduto(p);
+   }
 
-    @DeleteMapping("/deletebyprod/{id}")
-    public ResponseEntity<?> deleteIMGPROD(@PathVariable Long id) {
-        Produto p = pr.findById(id).orElseThrow(() -> new EntidadeException("Produto", id));
-        ir.deleteByProduto(p);
-        return ResponseEntity.status(HttpStatus.OK).body("Imagem deletada com sucesso!");
-    }
+   @DeleteMapping("/deletebyprod/{id}")
+   public ResponseEntity<?> deleteIMGPROD(@PathVariable Long id, HttpServletRequest request) {
+      if (!uc.checkAdmin(request)) {
+         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Acesso negado!");
+      } else {
+         Produto p = pr.findById(id).orElseThrow(() -> new EntidadeException("Produto", id));
+         ir.deleteByProduto(p);
+         return ResponseEntity.status(HttpStatus.OK).body("Imagem deletada com sucesso!");
+      }
+   }
 
 }
